(Note that this commit is very much alpha quality, with TODOs, missing javadocs, etc.  It is intended to be a proof-of-concept more than anything, and to double-check that I'm going down the right path.)
This first iteration supports injecting a Couchbase Cluster object.
This allows opening Couchbase resources such as Buckets and Collections,
and these in turn used for all data operations.

Basic integration tests are also provided.  These require a running
Couchbase cluster with some default configuration setup - see
the DefaultTest javadocs for details.

The configuration for the cluster is currently hardcoded, with the next
step being to load this from application.yml.